requirements.txt
Flask==1.0.2
psycopg2-binary==2.7.7
configparser==3.7.3